LLM-RAG-Toolkit/requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0

# OpenAI API
openai==0.27.8

# Document processing
PyMuPDF==1.22.3
pdfplumber==0.9.0
python-docx==0.8.11
beautifulsoup4==4.10.0

# Natural Language Processing
nltk==3.6.5
sentence-transformers==2.2.2
transformers==4.15.0

# Data handling and analysis
numpy==1.23.5
pandas==1.3.4

# Machine Learning
scikit-learn==1.0.1
torch==1.10.0

# Web framework (for potential future use)
streamlit==1.22.0

# Progress bars
tqdm==4.65.0

# Visualization (for potential future use)
matplotlib==3.5.0
seaborn==0.11.2

# Testing
pytest==6.2.5

# Linting and formatting
flake8==4.0.1
black==21.12b0

# Logging
rich==10.16.1

# Type checking
mypy==0.910

# CLI
click==8.0.3
